Categorise birthweight centiles and gestational ages into small vulnerable newborn strata

Categorise birthweight centiles and gestational ages into small vulnerable newborn strata

Usage

categorise_svn(p, gest_days)

Arguments

p

A numeric vector with birth weight centiles from the INTERGROWTH-21st Newborn Size standard. If elements of p are not between 0 and 1, gigs will warn you. You can customise this behaviour with the GIGS package-level options.

gest_days

A numeric vector of length one or more with gestational age(s) in days. Used to determine whether infants are term or preterm (a gestational age > 259 days means an infant is term).

Value

An object of class factor with the same length as the longest input vector, containing small vulnerable newborn classifications. Its levels are c("Preterm SGA", "Preterm AGA", "Preterm LGA", "Term SGA", "Term AGA", "Term LGA").

Details

Cut-offs for small, vulnerable newborn categorisations are:

SVN category	Factor level	Gestational age range	Centile bounds

Preterm-SGA	"Preterm SGA"	p < 0.1	gest_days < 259

Preterm-AGA	"Preterm AGA"	0.1 =< p =< 0.9	gest_days < 259

Preterm-LGA	"Preterm LGA"	p > 0.9	gest_days < 259

Term-SGA	"Term SGA"	p < 0.1	gest_days >= 259

Term-AGA	"Term AGA"	0.1 =< p =< 0.9	gest_days >= 259

Term-LGA	"Term LGA"	p > 0.9	gest_days >= 259

Abbreviations: SGA, small-for-gestational age; AGA, appropriate-for-gestational age; LGA, large-for-gestational age.

Note

Input vectors are recycled by vctrs::vec_recycle_common(), and must adhere to the vctrs recycling rules.

References

Lawn JE, Ohuma EO, Bradley E, Idueta LS, Hazel E, Okwaraji YB et al. Small babies, big risks: global estimates of prevalence and mortality for vulnerable newborns to accelerate change and improve counting. The Lancet 2023, 401(10389):1707-1719. doi:10.1016/S0140-6736(23)00522-6

See also

Examples

p <- c(0.01, 0.07, 0.25, 0.75, 0.93, 0.99)
gest_days <- c(250, 250, 250, 280, 280, 280) # 3 preterm, 3 term
categorise_svn(p = p, gest_days = gest_days)
#> ! Unused factor levels kept after small vulnerable newborn categorisation:
#> "Preterm LGA" and "Term SGA".
#> [1] Preterm SGA Preterm SGA Preterm AGA Term AGA    Term LGA    Term LGA   
#> Levels: Preterm SGA Preterm AGA Preterm LGA Term SGA Term AGA Term LGA