Classify size for gestational age using the INTERGROWTH-21st weight-for-gestational age standard

Size for gestational age categories are split by centile: small-for-GA (SGA; 10th centile), appropriate-for-GA (AGA; 10th to 90th centile) and large-for-GA (LGA; >90th centile). This function also supports classification of severe SGA (<3rd centile) using the severe parameter.

Usage

classify_sfga(weight_kg, gest_days, sex, severe = FALSE)

Arguments

weight_kg

Numeric vector of length one or more with weight value(s) in kg.

gest_days

Numeric vector of length one or more with gestational age(s) at birth in days between 168 and 300. By default, gigs will replace out-of-bounds elements in gest_days with NA and warn you. This behaviour can be customised using the functions in gigs_options.

sex

Character vector of length one or more with sex(es), either "M" (male) or "F" (female). This argument is case-sensitive. By default, gigs will replace elements of sex which are not "M" or "F" with NA and warn you. This behaviour can be customised using the functions in gigs_options.

severe

A single logical value specifying whether to categorise SGA values are below the third centile as "SGA(<3)". Default = FALSE.

Value

An object of class factor with the same length as the longest input vector, containing size-for-GA classifications. If severe = FALSE, levels are c("SGA", "AGA", "LGA"). If severe = TRUE, levels are c("SGA(<3)", "SGA", "AGA", "LGA").

Note

Input vectors are recycled by vctrs::vec_recycle_common(), and must adhere to the vctrs recycling rules.

References

WHO. Physical status: the use and interpretation of anthropometry. Report of a WHO Expert Committee. World Health Organisation Technical Report Series 1995, 854: 1–452

Royal College of Obstetricians and Gynaecologists. The Investigation and Management of the Small-for-Gestational-Age Fetus: Green-top Guideline No. 31. Technical report, Royal College of Obstetricians and Gynaecologists, London, 2013.

See also

Examples

# By default, does not differentiate between p < 0.03 and p < 0.10
classify_sfga(
  weight_kg = c(2.2, 3.3, 4.2),
  gest_days = 267:269,
  sex = c("M", "F", "M")
)
#> [1] SGA AGA LGA
#> Levels: SGA AGA LGA

# With severe = TRUE, highlights p < 0.03
classify_sfga(
  weight_kg = c(2.2, 3.3, 4.2),
  gest_days = 267:269,
  sex = c("M", "F", "M"),
  severe = TRUE
)
#> [1] SGA(<3) AGA     LGA    
#> Levels: SGA(<3) SGA AGA LGA