Package-level gigs options

An environment containing six named character vectors. These define how gigs handles inputs with missing, undefined, or invalid elements:

"handle_missing_data" - How should gigs handle missing (NA) elements?

"handle_undefined_data" - How should gigs handle undefined (NaN, Inf, -Inf) elements?

"handle_oob_xvar" - How should gigs handle elements of x variables which are out of bounds for the standard in use?

"handle_invalid_sex" - How should gigs handle elements of sex which are not one of "M" or "F"?

"handle_oob_centiles" - In centile2value functions, how should gigs handle elements of p that are not between 0 and 1?

Each of these options can take one of three values:

"quiet" - Invalid elements are set to NA, silently.

"warn" - Invalid elements are set to NA, with warnings issued when this is done.

"error" - Invalid elements will cause informative errors.

By default, each option in .gigs_options is set to "warn", so you will be informed of any invalid data used as input. Use gigs_option_set() or gigs_option_set() to change this behaviour.

Value

A named environment, where each name maps onto a specific option for the GIGS package.

See also

Examples

# Get the names of all available options
names(.gigs_options)
#> [1] "handle_oob_centiles"   "handle_unused_levels"  "handle_undefined_data"
#> [4] "handle_oob_xvar"       "handle_missing_data"   "handle_invalid_sex"