Get and set gigs package options

By default, gigs will handle invalid input data, e.g. missing (NA) or undefined (NaN, Inf and -Inf) values, by warning you of their presence and replacing the invalid values with NA. You can change this behaviour with gigs_option_set() so that gigs does this silently, or make gigs throw errors when confronted with bad data.

Usage

gigs_option_get(option, silent = FALSE)

gigs_option_set(option, new_value, silent = FALSE)

Arguments

option

A single character variable, one of:

"handle_missing" - How should gigs handle missing (NA) data?

"handle_undefined" - How should gigs handle undefined (NaN, Inf, -Inf) data?

"handle_invalid_xvar" - How should gigs handle x variables that are out of bounds for the standard in use?

"handle_invalid_sex" - How should gigs handle elements of sex which are not one of "M" or "F"?

"handle_invalid_acronym" - In conversion functions, how should gigs handle acronym elements which do not map onto any implemented growth standard?

"handle_invalid_centiles" - In centile2value functions, how should gigs handle elements of p that are not between 0 and 1?

silent

A single logical value denoting whether the function will send a message() to the console describing either:

The current value of option for gigs_option_get().

The newly-set value of option for gigs_option_set().

new_value

A single character variable, one of:

"quiet" - gigs will replace invalid vector elements to NA quietly

"warn" - gigs will replace invalid vector elements with NA loudly

"error" - gigs will throw informative errors if invalid inputs are encountered

Value

For gigs_option_get(), noisiliy returns the current value of option. For gigs_option_set(), invisibly returns the new value of option.

Examples

if (FALSE) {
  library(gigs)

  # Get the names of all available options
  names(.gigs_options)

  # Retrieve the value of a given option --> will invisibly return value and
  # print to console
  gigs_option_get("handle_missing_data")
  # Set the value of an option --> will invisibly return the new value and
  # print your change to the console
  gigs_option_set("handle_missing_data", "quiet")
}