Convert values to z-scores/centiles in the INTERGROWTH-21st Fetal Growth standards

Convert values to z-scores/centiles in the INTERGROWTH-21st Fetal Growth standards

Usage

ig_fet_value2zscore(y, x, acronym)

ig_fet_hcfga_value2zscore(headcirc_mm, gest_days)

ig_fet_bpdfga_value2zscore(bpd_mm, gest_days)

ig_fet_acfga_value2zscore(abdocirc_mm, gest_days)

ig_fet_flfga_value2zscore(femurlen_mm, gest_days)

ig_fet_ofdfga_value2zscore(ofd_mm, gest_days)

ig_fet_efwfga_value2zscore(efw_g, gest_days)

ig_fet_sfhfga_value2zscore(sfh_cm, gest_days)

ig_fet_crlfga_value2zscore(crl_mm, gest_days)

ig_fet_gafcrl_value2zscore(gest_days, crl_mm)

ig_fet_gwgfga_value2zscore(gest_wt_gain_kg, gest_days)

ig_fet_pifga_value2zscore(puls_idx, gest_days)

ig_fet_rifga_value2zscore(resist_idx, gest_days)

ig_fet_sdrfga_value2zscore(sys_dia_ratio, gest_days)

ig_fet_tcdfga_value2zscore(tcd_mm, gest_days)

ig_fet_gaftcd_value2zscore(gest_days, tcd_mm)

ig_fet_poffga_value2zscore(par_occ_fiss_mm, gest_days)

ig_fet_sffga_value2zscore(sylv_fiss_mm, gest_days)

ig_fet_avfga_value2zscore(ant_hlv_mm, gest_days)

ig_fet_pvfga_value2zscore(atr_phlv_mm, gest_days)

ig_fet_cmfga_value2zscore(cist_mag_mm, gest_days)

ig_fet_value2centile(y, x, acronym)

ig_fet_hcfga_value2centile(headcirc_mm, gest_days)

ig_fet_bpdfga_value2centile(bpd_mm, gest_days)

ig_fet_acfga_value2centile(abdocirc_mm, gest_days)

ig_fet_flfga_value2centile(femurlen_mm, gest_days)

ig_fet_ofdfga_value2centile(ofd_mm, gest_days)

ig_fet_efwfga_value2centile(efw_g, gest_days)

ig_fet_sfhfga_value2centile(sfh_cm, gest_days)

ig_fet_crlfga_value2centile(crl_mm, gest_days)

ig_fet_gafcrl_value2centile(gest_days, crl_mm)

ig_fet_gwgfga_value2centile(gest_wt_gain_kg, gest_days)

ig_fet_pifga_value2centile(puls_idx, gest_days)

ig_fet_rifga_value2centile(resist_idx, gest_days)

ig_fet_sdrfga_value2centile(sys_dia_ratio, gest_days)

ig_fet_tcdfga_value2centile(tcd_mm, gest_days)

ig_fet_gaftcd_value2centile(gest_days, tcd_mm)

ig_fet_poffga_value2centile(par_occ_fiss_mm, gest_days)

ig_fet_sffga_value2centile(sylv_fiss_mm, gest_days)

ig_fet_avfga_value2centile(ant_hlv_mm, gest_days)

ig_fet_pvfga_value2centile(atr_phlv_mm, gest_days)

ig_fet_cmfga_value2centile(cist_mag_mm, gest_days)

Arguments

y

Numeric vector of length one or more with anthropometric measurement(s) to convert to centiles/z-scores. Units depend on which acronym(s) are in use.

x

Numeric vector of length one or more with x values. Elements should have specific units and be between certain values depending on the standard in use (defined by acronym). These are:

Between 98 and 280 days for "hcfga", "bpdfga", "acfga", "flfga", "ofdfga", and "tcdfga".

Between 154 and 280 days for "efwfga".

Between 112 and 294 days for "sfhfga".

Between 58 and 105 days for "crlfga".

Between 19 and 95 mm for "gafcrl".

Between 105 and 280 days for "gwgfga".

Between 168 and 280 days for "pifga", "rifga", and "sdrfga" (the INTERGROWTH-21st Fetal Doppler standards).

Between 105 and 252 days for "poffga", "sffga", "avfga", and "pvfga" (the INTERGROWTH-21st Fetal Brain Development standards).

By default, gigs will replace out-of-bounds elements in x with NA and warn you. This behaviour can be customised using the functions in gigs_options.

acronym

Character vector of length one or more denoting the INTERGROWTH-21st Fetal Growth standard(s) in use. Each value should be one of:

"hcfga" (head circumference-for-GA)

"bpdfga" (biparietal diameter-for-GA)

"acfga" (abdominal circumference-for-GA)

"flfga" (femur length-for-GA)

"ofdfga" (occipito-frontal diameter-for-GA)

"efwfga" (estimated fetal weight-for-GA)

"sfhfga" (symphisis-fundal height-for-GA)

"crlfga" (crown-rump length-for-GA)

"gafcrl" (GA-for-crown-rump length)

"gwgfga" (gestational weight gain-for-GA)

"pifga" (pulsatility index-for-GA)

"rifga" (resistance index-for-GA)

"sdrfga" (systolic/diastolic ratio-for-GA)

"tcdfga" (transcerebellar diameter-for-GA)

"poffga" (parieto-occipital fissure-for-GA)

"sffga" (Sylvian fissure-for-GA)

"avfga" (anterior horn of lateral ventricle-for-GA)

"pvfga" (atrium of posterior horn of lateral ventricle-for-GA)

"cmfga" (cisterna magna-for-GA)

This argument is case-sensitive. By default, gigs will replace elements in acronym which are not one of the above values with NA and warn you. This behaviour can be customised using the functions in gigs_options. If all elements in acronym are not one of the above values, gigs will throw an error.

headcirc_mm

Numeric vector of length one or more with head circumference measurements in mm.

gest_days

Numeric vector of length one or more that is usually used as the x variable, but can also be a y variable for the "gafcrl" and "gaftcd" standards. When gest_days is being used as the x variable, out of bounds elements of gest_days will be handled according to the gigs_options for "handle_invalid_xvar".

bpd_mm

Numeric vector of length one or more with biparietal diameter measurement(s) in mm.

abdocirc_mm

Numeric vector of length one or more with abdominal circumference measurement(s) in mm.

femurlen_mm

Numeric vector of length one or more with femur length(s) in mm.

ofd_mm

Numeric vector of length one or more with in occipito-frontal diameter(s) in mm.

efw_g

Numeric vector of length one or more with estimated fetal weight(s) in g.

sfh_cm

Numeric vector of length one or more with symphisis-fundal height(s) in mm.

crl_mm

Numeric vector of length one or more with crown-rump length(s) in mm.

gest_wt_gain_kg

Numeric vector of length one or more with gestational weight gain(s)/loss(es) in kg.

puls_idx

Numeric vector of length one or more with pulsatility index value(s).

resist_idx

Numeric vector of length one or more with resistance index value(s).

sys_dia_ratio

Numeric vector of length one or more with systolic/diastolic ratio value(s).

tcd_mm

Numeric vector of length one or more with transcerebellar diameter(s) in mm.

par_occ_fiss_mm

Numeric vector of length one or more with parietal-occipital fissure measurement(s) in mm.

sylv_fiss_mm

Numeric vector of length one or more with Sylvian fissure measurement(s) in mm.

ant_hlv_mm

Numeric vector of length one or more with anterior horn of lateral ventricle measurements(s) in mm.

atr_phlv_mm

Numeric vector of length one or more with atrium of posterior horn of lateral ventricle measurement(s) in mm.

cist_mag_mm

Numeric vector of length one or more with cisterna magna measurement(s) in mm.

Value

Numeric vector of z-scores/centiles with length equal to the longest input vector.

References

Papageorghiou AT, Ohuma EO, Altman DG, Todros T, Cheikh Ismail L, Lambert A et al. International standards for fetal growth based on serial ultrasound measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project. Lancet 2014, 384(9946):869-79. doi:10.1016/S0140-6736(14)61490-2

Stirnemann J, Villar J, Salomon LJ, Ohuma EO, Lamber A, Victoria CG et al. International Estimated Fetal Weight Standards of the INTERGROWTH-21st Project. Ultrasound Obstet Gynecol 2016, 49:478-486 doi:10.1002/uog.17347

Papageorghiou AT, Ohuma EO, Gravett MG, Lamber A, Noble JA, Pang R et al. International standards for symphysis-fundal height based on serial measurements from the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project: prospective cohort study in eight countries. BMJ 2016, 355:i5662 doi:10.1136/bmj.i5662

Papageorghiou AT, Kennedy SH, Salomon LJ, Ohuma EO, Cheikh Ismail L, Barros FC et al. International standards for early fetal size and pregnancy dating based on ultrasound measurement of crown-rump length in the first trimester of pregnancy. Ultrasound Obstet Gynecol 2014, 44(6):641-48 doi:10.1002/uog.13448

Cheikh Ismail L, Bishop DC, Pang R, Ohuma EO, Kac G, Abrams B et al. Gestational weight gain standards based on women enrolled in the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project: a prospective longitudinal cohort study. BMJ 2016, 352:i555 doi:10.1136/bmj.i555

Drukker L, Staines-Urias E, Villar J, Barros FC, Carvalho M, Munim S et al. International gestational age-specific centiles for umbilical artery Doppler indices: a longitudinal prospective cohort study of the INTERGROWTH-21st Project. Am J Obstet Gynecol 2021, 222(6):602.e1-602.e15 doi:10.1016/j.ajog.2020.01.012

Rodriguez-Sibaja MJ, Villar J, Ohuma EO, Napolitano R, Heyl S, Carvalho M et al. Fetal cerebellar growth and Sylvian fissure maturation: international standards from Fetal Growth Longitudinal Study of INTERGROWTH-21st Project Ultrasound Obstet Gynecol 2021, 57(4):614-623 doi:10.1002/uog.22017

Napolitano R, Molloholli M, Donadono V, Ohuma EO, Wanyonyi SZ, Kemp B et al. International standards for fetal brain structures based on serial ultrasound measurements from Fetal Growth Longitudinal Study of INTERGROWTH-21st Project Ultrasound Obstet Gynecol 2020, 56(3):359-370 doi:10.1002/uog.21990

Examples

# Convert values to centiles
y <- 335
ig_fet_value2centile(y = y, x = 280, acronym = "hcfga") |>
  round(digits = 2)
#> [1] 0.53

# Or values to z-scores
ig_fet_value2zscore(y = y, x = 280, acronym = "hcfga") |>
  round(digits = 2)
#> [1] 0.08

# Specify which standard to use with the acronym parameter...
ig_fet_value2zscore(y = 340, x = 280, acronym = "acfga") |>
  round(digits = 2)
#> [1] -0.44

# ... or by using a standard-specific function
ig_fet_acfga_value2zscore(abdocirc_mm = 340, gest_days = 280) |>
  round(digits = 2)
#> [1] -0.44

# Inputs are recycled to the input of the longest length
ig_fet_efwfga_value2zscore(efw_g = seq(3200, 3500, by = 50),
                           gest_days = 280) |>
  round(digits = 2)
#> [1] -0.34 -0.22 -0.09  0.03  0.15  0.28  0.40

# Bad inputs will not stop the function but will instead return NA - here 90
# days for `gest_days` is outside the bounds of the INTERGROWTH-21st Fetal
# Growth standards
ig_fet_hcfga_value2zscore(headcirc_mm = c(90, 100, 300, 310),
                          gest_days = c(91, 98, 224, 266)) |>
  round(digits = 2)
#> Warning: Variable 'gest_days': 1 in 4 elements were out-of-bounds (see the 'ig_fet' conversion functions documentation).
#> [1]    NA  0.38  0.59 -1.55