Convert z-scores/centiles to values in the INTERGROWTH-21st Fetal standards

Convert z-scores/centiles to values in the INTERGROWTH-21st Fetal standards

Usage

ig_fet_zscore2value(z, x, acronym)

ig_fet_hcfga_zscore2value(z, gest_days)

ig_fet_bpdfga_zscore2value(z, gest_days)

ig_fet_acfga_zscore2value(z, gest_days)

ig_fet_flfga_zscore2value(z, gest_days)

ig_fet_ofdfga_zscore2value(z, gest_days)

ig_fet_efwfga_zscore2value(z, gest_days)

ig_fet_sfhfga_zscore2value(z, gest_days)

ig_fet_crlfga_zscore2value(z, gest_days)

ig_fet_gafcrl_zscore2value(z, crl_mm)

ig_fet_gwgfga_zscore2value(z, gest_days)

ig_fet_pifga_zscore2value(z, gest_days)

ig_fet_rifga_zscore2value(z, gest_days)

ig_fet_sdrfga_zscore2value(z, gest_days)

ig_fet_tcdfga_zscore2value(z, gest_days)

ig_fet_gaftcd_zscore2value(z, tcd_mm)

ig_fet_poffga_zscore2value(z, gest_days)

ig_fet_sffga_zscore2value(z, gest_days)

ig_fet_avfga_zscore2value(z, gest_days)

ig_fet_pvfga_zscore2value(z, gest_days)

ig_fet_cmfga_zscore2value(z, gest_days)

ig_fet_centile2value(p, x, acronym)

ig_fet_hcfga_centile2value(p, gest_days)

ig_fet_bpdfga_centile2value(p, gest_days)

ig_fet_acfga_centile2value(p, gest_days)

ig_fet_flfga_centile2value(p, gest_days)

ig_fet_ofdfga_centile2value(p, gest_days)

ig_fet_efwfga_centile2value(p, gest_days)

ig_fet_sfhfga_centile2value(p, gest_days)

ig_fet_crlfga_centile2value(p, gest_days)

ig_fet_gafcrl_centile2value(p, crl_mm)

ig_fet_gwgfga_centile2value(p, gest_days)

ig_fet_pifga_centile2value(p, gest_days)

ig_fet_rifga_centile2value(p, gest_days)

ig_fet_sdrfga_centile2value(p, gest_days)

ig_fet_tcdfga_centile2value(p, gest_days)

ig_fet_gaftcd_centile2value(p, tcd_mm)

ig_fet_poffga_centile2value(p, gest_days)

ig_fet_sffga_centile2value(p, gest_days)

ig_fet_avfga_centile2value(p, gest_days)

ig_fet_pvfga_centile2value(p, gest_days)

ig_fet_cmfga_centile2value(p, gest_days)

Arguments

z, p

Numeric vector of length one or more with centiles/z-scores to convert to values. For p, if an element of p is not between 0 and 1, gigs will replace it with NA and warn you. This behaviour can be customised using the functions in gigs_options.

x

Numeric vector of length one or more with x values. Elements should have specific units and be between certain values depending on the standard in use (defined by acronym). These are:

Between 98 and 280 days for "hcfga", "bpdfga", "acfga", "flfga", "ofdfga", and "tcdfga".

Between 154 and 280 days for "efwfga".

Between 112 and 294 days for "sfhfga".

Between 58 and 105 days for "crlfga".

Between 19 and 95 mm for "gafcrl".

Between 105 and 280 days for "gwgfga".

Between 168 and 280 days for "pifga", "rifga", and "sdrfga" (the INTERGROWTH-21st Fetal Doppler standards).

Between 105 and 252 days for "poffga", "sffga", "avfga", and "pvfga" (the INTERGROWTH-21st Fetal Brain Development standards).

By default, gigs will replace out-of-bounds elements in x with NA and warn you. This behaviour can be customised using the functions in gigs_options.

acronym

Character vector of length one or more denoting the INTERGROWTH-21st Fetal Growth standard(s) in use. Each value should be one of:

"hcfga" (head circumference-for-GA)

"bpdfga" (biparietal diameter-for-GA)

"acfga" (abdominal circumference-for-GA)

"flfga" (femur length-for-GA)

"ofdfga" (occipito-frontal diameter-for-GA)

"efwfga" (estimated fetal weight-for-GA)

"sfhfga" (symphisis-fundal height-for-GA)

"crlfga" (crown-rump length-for-GA)

"gafcrl" (GA-for-crown-rump length)

"gwgfga" (gestational weight gain-for-GA)

"pifga" (pulsatility index-for-GA)

"rifga" (resistance index-for-GA)

"sdrfga" (systolic/diastolic ratio-for-GA)

"tcdfga" (transcerebellar diameter-for-GA)

"poffga" (parieto-occipital fissure-for-GA)

"sffga" (Sylvian fissure-for-GA)

"avfga" (anterior horn of lateral ventricle-for-GA)

"pvfga" (atrium of posterior horn of lateral ventricle-for-GA)

"cmfga" (cisterna magna-for-GA)

This argument is case-sensitive. By default, gigs will replace elements in acronym which are not one of the above values with NA and warn you. This behaviour can be customised using the functions in gigs_options. If all elements in acronym are not one of the above values, gigs will throw an error.

gest_days, crl_mm, tcd_mm

Numeric vector of length one or more with standard-specific x variables. See the documentation for x for information on how out-of-bounds elements will be handled.

Value

Numeric vector of expected measurements with length equal to the longest input vector.

Note

Input vectors are recycled by vctrs::vec_recycle_common(), and must adhere to the vctrs recycling rules.

References

Papageorghiou AT, Ohuma EO, Altman DG, Todros T, Cheikh Ismail L, Lambert A et al. International standards for fetal growth based on serial ultrasound measurements: the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project. Lancet 2014, 384(9946):869-79. doi:10.1016/S0140-6736(14)61490-2

Stirnemann J, Villar J, Salomon LJ, Ohuma EO, Lamber A, Victoria CG et al. International Estimated Fetal Weight Standards of the INTERGROWTH-21st Project. Ultrasound Obstet Gynecol 2016, 49:478-486 doi:10.1002/uog.17347

Papageorghiou AT, Ohuma EO, Gravett MG, Lamber A, Noble JA, Pang R et al. International standards for symphysis-fundal height based on serial measurements from the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project: prospective cohort study in eight countries. BMJ 2016, 355:i5662 doi:10.1136/bmj.i5662

Papageorghiou AT, Kennedy SH, Salomon LJ, Ohuma EO, Cheikh Ismail L, Barros FC et al. International standards for early fetal size and pregnancy dating based on ultrasound measurement of crown-rump length in the first trimester of pregnancy. Ultrasound Obstet Gynecol 2014, 44(6):641-48 doi:10.1002/uog.13448

Cheikh Ismail L, Bishop DC, Pang R, Ohuma EO, Kac G, Abrams B et al. Gestational weight gain standards based on women enrolled in the Fetal Growth Longitudinal Study of the INTERGROWTH-21st Project: a prospective longitudinal cohort study. BMJ 2016, 352:i555 doi:10.1136/bmj.i555

Drukker L, Staines-Urias E, Villar J, Barros FC, Carvalho M, Munim S et al. International gestational age-specific centiles for umbilical artery Doppler indices: a longitudinal prospective cohort study of the INTERGROWTH-21st Project. Am J Obstet Gynecol 2021, 222(6):602.e1-602.e15 doi:10.1016/j.ajog.2020.01.012

Rodriguez-Sibaja MJ, Villar J, Ohuma EO, Napolitano R, Heyl S, Carvalho M et al. Fetal cerebellar growth and Sylvian fissure maturation: international standards from Fetal Growth Longitudinal Study of INTERGROWTH-21st Project Ultrasound Obstet Gynecol 2021, 57(4):614-623 doi:10.1002/uog.22017

Napolitano R, Molloholli M, Donadono V, Ohuma EO, Wanyonyi SZ, Kemp B et al. International standards for fetal brain structures based on serial ultrasound measurements from Fetal Growth Longitudinal Study of INTERGROWTH-21st Project Ultrasound Obstet Gynecol 2020, 56(3):359-370 doi:10.1002/uog.21990

Examples

# Convert centiles to values
ig_fet_centile2value(p = c(0.25, 0.5, 0.75), # 25th, 50th, 75th centile
                     x = 279:281,
                     acronym = "hcfga") |>
  round(digits = 2)
#> Warning: Variable 'x': 1 in 3 elements were out-of-bounds (see the 'ig_fet' conversion functions documentation).
#> [1] 324.98 333.95     NA

# Or z-scores to values
ig_fet_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                    x = 279:281,
                    acronym = "hcfga") |>
  round(digits = 2)
#> Warning: Variable 'x': 1 in 3 elements were out-of-bounds (see the 'ig_fet' conversion functions documentation).
#> [1] 324.98 333.95     NA

# Specify which standard to use with the acronym parameter...
ig_fet_zscore2value(z = -1:1,
                    x = 279:281,
                    acronym = "acfga") |>
  round(digits = 2)
#> Warning: Variable 'x': 1 in 3 elements were out-of-bounds (see the 'ig_fet' conversion functions documentation).
#> [1] 326.34 349.80     NA

# ... or by using a standard-specific function
ig_fet_acfga_zscore2value(z = -1:1,
                          gest_days = 279:281) |>
  round(digits = 2)
#> Warning: Variable 'gest_days': 1 in 3 elements were out-of-bounds (see the 'ig_fet' conversion functions documentation).
#> [1] 326.34 349.80     NA

# Inputs are recycled to the input of the longest length
ig_fet_acfga_zscore2value(z = 0,
                          gest_days = 279:281) |>
  round(digits = 2)
#> Warning: Variable 'gest_days': 1 in 3 elements were out-of-bounds (see the 'ig_fet' conversion functions documentation).
#> [1] 348.52 349.80     NA

# Bad inputs will not stop the function but will instead produce `NA`s in the
# output - by default gigs will issue useful warnings
ig_fet_hcfga_centile2value(p = c(-0.5, 0.15, 0.5, 0.85),
                           gest_days = c(91, 98, 224, NA)) |>
  round(digits = 2)
#> Warning: Variable 'p': 1 in 4 elements were not between 0 and 1.
#> Warning: Variable 'gest_days': 1 in 4 elements were missing (`NA`).
#> Warning: Variable 'gest_days': 1 in 4 elements were out-of-bounds (see the 'ig_fet' conversion functions documentation).
#> [1]     NA  92.09 294.37     NA