Convert values to z-scores/percentiles in the INTERGROWTH-21st Newborn Size Standards

Convert values to z-scores/percentiles in the INTERGROWTH-21st Newborn Size Standards

Usage

ig_nbs_value2percentile(y, gest_age, sex, acronym)

ig_nbs_wfga_value2percentile(weight_kg, gest_age, sex)

ig_nbs_lfga_value2percentile(length_cm, gest_age, sex)

ig_nbs_hcfga_value2percentile(headcirc_cm, gest_age, sex)

ig_nbs_wlrfga_value2percentile(wei_len_ratio, gest_age, sex)

ig_nbs_fmfga_value2percentile(fat_mass_g, gest_age, sex)

ig_nbs_bfpfga_value2percentile(body_fat_perc, gest_age, sex)

ig_nbs_ffmfga_value2percentile(fatfree_mass_g, gest_age, sex)

ig_nbs_value2zscore(y, gest_age, sex, acronym)

ig_nbs_wfga_value2zscore(weight_kg, gest_age, sex)

ig_nbs_lfga_value2zscore(length_cm, gest_age, sex)

ig_nbs_hcfga_value2zscore(headcirc_cm, gest_age, sex)

ig_nbs_wlrfga_value2zscore(wei_len_ratio, gest_age, sex)

ig_nbs_fmfga_value2zscore(fat_mass_g, gest_age, sex)

ig_nbs_bfpfga_value2zscore(body_fat_perc, gest_age, sex)

ig_nbs_ffmfga_value2zscore(fatfree_mass_g, gest_age, sex)

Arguments

y

Value(s) to convert to either a z-score/z-scores or a percentile/percentiles.

gest_age

Gestational age(s) in days. Must be between 266 to 294 for body composition equations ("fmfga", "bfpfga", or "ffmfga"), or between 168 and 300 for the other standards.

sex

Sex(es), either "M" (male) or "F" (female).

acronym

Acronym(s) denoting the INTERGROWTH-21st NBS standard to use. Must be one of "wfga" (weight-for-GA), "lfga" (length-for-GA), "hcfga" (head circumference-for-GA), "wlrfga" (weight/length ratio-for-GA), "fmfga" (fat mass-for-GA), "bfpfga" (body fat %-for-GA), or "ffmfga" (fat-free mass-for-GA).

weight_kg

Birth weight(s) in kg

length_cm

Birth length(s) in cm

headcirc_cm

Birth head circumference(s) in cm

wei_len_ratio

Weight-length ratio(s) in kg per m.

fat_mass_g

Fat mass(es) in g

body_fat_perc

Body fat percentage(s)

fatfree_mass_g

Fat-free mass(es) in g

Value

Z-scores/percentiles for each combination of measurement, gestational age, sex, and acronym provided to the function.

Note

Input vectors will be recycled to the length of the longest vector according to the rules of vctrs::vec_recycle_common().

References

Villar J, Cheikh Ismail L, Victora CG, Ohuma EO, Bertino E, Altman DG, et al. International standards for newborn weight, length, and head circumference by gestational age and sex: the Newborn Cross-Sectional Study of the INTERGROWTH-21st Project. Lancet 2014, 384(9946):857-68. doi:10.1016/S0140-6736(14)60932-6

Villar J, Giuliani F, Fenton TR, Ohuma EO, Ismail LC, Kennedy SH et al. INTERGROWTH-21st very preterm size at birth reference charts. Lancet 2016, 387(10021):844-45. doi:10.1016/S0140-6736(16)00384-6

Villar J, Puglia FA, Fenton TR, Ismal LC, Staines-Urias E, Giuliani F, et al. Body composition at birth and its relationship with neonatal anthropometric ratios: the newborn body composition study of the INTERGROWTH-21st project. Pediatric Research 2017, 82:305-316. doi:10.1038/pr.2017.52

Examples

# Convert values to percentiles
ig_nbs_value2percentile(y = 3.12, gest_age = 280, sex = "M", acronym = "wfga") |>
  round(digits = 2)
#> [1] 0.25

# Or values to z-scores
ig_nbs_value2zscore(y = 3.12, gest_age = 280, sex = "M", acronym = "wfga") |>
  round(digits = 2)
#> [1] -0.67

# Specify which standard to use with the acronym parameter...
ig_nbs_value2zscore(y = 48.84, gest_age = 280, sex = "F", acronym = "lfga") |>
  round(digits = 2)
#> [1] -0.25

# ... or by using a standard-specific function
ig_nbs_lfga_value2zscore(length_cm = 48.84, gest_age = 280, sex = "F") |>
  round(digits = 2)
#> [1] -0.25

# Inputs are recycled to the input of the longest length
ig_nbs_wlrfga_value2zscore(wei_len_ratio = c(7.37, 6.47, 6.12, 6.86),
                           gest_age = 280,
                           sex = "M") |>
  round(digits = 2)
#> [1]  0.85 -0.46 -0.97  0.11

# Bad inputs will not stop the function but will instead return NA - here 301
# days gest_age is outside the bounds of the INTERGROWTH-21st newborn size
# standards
ig_nbs_hcfga_value2percentile(headcirc_cm = c(23.0, 28.0, 33.0, 35.0),
                              gest_age = c(168, 217, 266, 301),
                              sex = "F") |>
  round(digits = 2)
#> [1] 0.72 0.42 0.49   NA