Convert values to z-scores/centiles in the INTERGROWTH-21st Postnatal Growth Standards for preterm infants

Convert values to z-scores/centiles in the INTERGROWTH-21st Postnatal Growth Standards for preterm infants

Usage

ig_png_value2zscore(y, x, sex, acronym)

ig_png_wfa_value2zscore(weight_kg, pma_weeks, sex)

ig_png_lfa_value2zscore(length_cm, pma_weeks, sex)

ig_png_hcfa_value2zscore(headcirc_cm, pma_weeks, sex)

ig_png_wfl_value2zscore(weight_kg, length_cm, sex)

ig_png_value2centile(y, x, sex, acronym)

ig_png_wfa_value2centile(weight_kg, pma_weeks, sex)

ig_png_lfa_value2centile(length_cm, pma_weeks, sex)

ig_png_hcfa_value2centile(headcirc_cm, pma_weeks, sex)

ig_png_wfl_value2centile(weight_kg, length_cm, sex)

Arguments

y

Numeric vector of length one or more with anthropometric measurement(s) to convert to centiles/z-scores. Units depend on which acronym(s) are in use.

x

Numeric vector of length one or more with x values. Elements of x or its standard-specific equivalents (pma_weeks, length_cm) should have specific units and be between certain values depending on the standard in use (defined by acronym). These are:

Between 27 and 64 weeks for "wfa", "lfa", and "hcfa".

Between 35 and 65 days for "wfl".

By default, gigs will replace out-of-bounds elements in x with NA and warn you. This behaviour can be customised using the functions in gigs_options.

sex

Character vector of length one or more with sex(es), either "M" (male) or "F" (female). This argument is case-sensitive. By default, gigs will replace elements of sex which are not "M" or "F" with NA and warn you. This behaviour can be customised using the functions in gigs_options.

acronym

Character vector of length one or more denoting the INTERGROWTH-21st Postnatal Growth standard(s) in use. Each element should be one of:

"wfa" (weight-for-age)

"lfa" (length-for-age)

"hcfa" (head circumference-for-age)

"wfl" (weight-for-length)

This argument is case-sensitive. By default, gigs will replace elements in acronym which are not one of the above values with NA and warn you. This behaviour can be customised using the functions in gigs_options. If all elements in acronym are not one of the above values, gigs will throw an error.

weight_kg

Numeric vector of length one or more with weight measurement(s) in kg.

pma_weeks

Numeric vector of length one or more with post-menstrual age(s) in weeks. Values should be within the bounds defined in the documentation for x.

length_cm

Numeric vector of length one or more with recumbent length measurement(s) in cm. This argument can be either an x variable when using the length-for-age standard ("lfa"), or a y variable when using the weight-for-length ("wfl") standard.

headcirc_cm

Numeric vector of length one or more with head circumference measurement(s) in cm.

Value

Numeric vector of z-scores/centiles with length equal to the longest input vector.

Note

Input vectors are recycled by vctrs::vec_recycle_common(), and must adhere to the vctrs recycling rules.

References

Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al. Postnatal growth standards for preterm infants: the Preterm Postnatal Follow-up Study of the INTERGROWTH-21st Project. Lancet Glob Health 2015, 3(11):e681-e691. doi:10.1016/S2214-109X(15)00163-1

Examples

# Convert values to centiles
ig_png_value2centile(y = c(5.79, 5.92, 7.25),
                     x = 54:56,
                     sex = c("M", "F", "M"),
                     acronym = "wfa") |>
  round(digits = 2)
#> [1] 0.25 0.50 0.75

# Or values to z-scores
ig_png_value2zscore(y = c(5.79, 5.92, 7.25),
                    x = 54:56,
                    sex = c("M", "F", "M"),
                    acronym = "wfa") |>
  round(digits = 2)
#> [1] -0.67  0.00  0.68

# Specify which standard to use with the acronym parameter...
ig_png_value2zscore(y = c(60.2, 60.3, 64.1),
                    x = 54:56,
                    sex = c("M", "F", "M"),
                    acronym = "lfa") |>
  round(digits = 2)
#> [1] -0.69  0.02  0.62

# ... or by using a standard-specific function
ig_png_lfa_value2zscore(length_cm = c(60.2, 60.3, 64.1),
                        pma_weeks = 54:56,
                        sex = c("M", "F", "M")) |>
  round(digits = 2)
#> [1] -0.69  0.02  0.62

# Inputs are recycled to the input of the longest length
ig_png_lfa_value2zscore(length_cm = c(51.10, 51.48, 51.86, 52.24, 52.63),
                        pma_weeks = 40,
                        sex = "M") |>
  round(digits = 2)
#> [1] 0.1 0.3 0.5 0.7 0.9

# Bad inputs will not stop the function but will instead produce `NA`s in the
# output - by default gigs will issue useful warnings
ig_png_hcfa_value2centile(headcirc_cm = c(20.6, NA, 38.2, 42.8),
                          pma_weeks = c(25, 27, 46, 64),
                          sex = "M") |>
  round(digits = 2)
#> Warning: Variable 'headcirc_cm': 1 in 4 elements were missing (`NA`).
#> Warning: Variable 'pma_weeks': 1 in 4 elements were out-of-bounds (see the 'ig_png' conversion functions documentation).
#> [1]   NA   NA 0.65 0.43