Convert z-scores/centiles to values in the INTERGROWTH-21st Postnatal Growth Standards for preterm infants

Convert z-scores/centiles to values in the INTERGROWTH-21st Postnatal Growth Standards for preterm infants

Usage

ig_png_zscore2value(z, x, sex, acronym)

ig_png_wfa_zscore2value(z, pma_weeks, sex)

ig_png_lfa_zscore2value(z, pma_weeks, sex)

ig_png_hcfa_zscore2value(z, pma_weeks, sex)

ig_png_wfl_zscore2value(z, length_cm, sex)

ig_png_centile2value(p, x, sex, acronym)

ig_png_wfa_centile2value(p, pma_weeks, sex)

ig_png_lfa_centile2value(p, pma_weeks, sex)

ig_png_hcfa_centile2value(p, pma_weeks, sex)

ig_png_wfl_centile2value(p, length_cm, sex)

Arguments

z, p

Numeric vector of length one or more with centiles/z-scores to convert to values. For p, if an element of p is not between 0 and 1, gigs will replace it with NA and warn you. This behaviour can be customised using the functions in gigs_options.

x

Numeric vector of length one or more with x values. Elements of x or its standard-specific equivalents (pma_weeks, length_cm) should have specific units and be between certain values depending on the standard in use (defined by acronym). These are:

Between 27 and 64 weeks for "wfa", "lfa", and "hcfa".

Between 35 and 65 days for "wfl".

By default, gigs will replace out-of-bounds elements in x with NA and warn you. This behaviour can be customised using the functions in gigs_options.

sex

Character vector of length one or more with sex(es), either "M" (male) or "F" (female). This argument is case-sensitive. By default, gigs will replace elements of sex which are not "M" or "F" with NA and warn you. This behaviour can be customised using the functions in gigs_options.

acronym

Character vector of length one or more denoting the INTERGROWTH-21st Postnatal Growth standard(s) in use. Each element should be one of:

"wfa" (weight-for-age)

"lfa" (length-for-age)

"hcfa" (head circumference-for-age)

"wfl" (weight-for-length)

This argument is case-sensitive. By default, gigs will replace elements in acronym which are not one of the above values with NA and warn you. This behaviour can be customised using the functions in gigs_options. If all elements in acronym are not one of the above values, gigs will throw an error.

pma_weeks, length_cm

Numeric vector of length one or more with standard-specific x variables. See the documentation for x for information the acceptable bounds of these variables, and on how out-of-bounds elements will be handled.

Value

Numeric vector of expected measurements with length equal to the longest input vector.

Note

Input vectors are recycled by vctrs::vec_recycle_common(), and must adhere to the vctrs recycling rules.

References

Villar J, Giuliani F, Bhutta ZA, Bertino E, Ohuma EO, Ismail LC et al. Postnatal growth standards for preterm infants: the Preterm Postnatal Follow-up Study of the INTERGROWTH-21st Project. Lancet Glob Health 2015, 3(11):e681-e691. doi:10.1016/S2214-109X(15)00163-1

Examples

# Convert centiles to values
ig_png_centile2value(p = c(0.25, 0.5, 0.75),
                     x = 54:56,
                     sex = c("M", "F", "M"),
                     acronym = "wfa") |>
  round(digits = 2)
#> [1] 5.79 5.92 7.25

# Or z-scores to values
ig_png_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                    x = 54:56,
                    sex = c("M", "F", "M"),
                    acronym = "wfa") |>
  round(digits = 2)
#> [1] 5.79 5.92 7.25

# Specify which standard to use with the acronym parameter...
ig_png_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                    x = 54:56,
                    sex = c("M", "F", "M"),
                    acronym = "lfa") |>
  round(digits = 2)
#> [1] 60.23 60.27 64.21

# ... or by using a standard-specific function
ig_png_lfa_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                        pma_weeks = 54:56,
                        sex = c("M", "F", "M")) |>
  round(digits = 2)
#> [1] 60.23 60.27 64.21

# Inputs are recycled to the input of the longest length
ig_png_lfa_zscore2value(z = seq(-0.5, 0.5, by = 0.2),
                        pma_weeks = 40,
                        sex = "M") |>
  round(digits = 2)
#> [1] 49.98 50.35 50.73 51.10 51.48 51.86

# Bad inputs will not stop the function but will instead produce `NA`s in the
# output - by default gigs will issue useful warnings
ig_png_hcfa_zscore2value(z = 0,
                         pma_weeks = c(25, 27, 46, 64),
                         sex = c("M", "M", NA, "M")) |>
  round(digits = 2)
#> Warning: Variable 'pma_weeks': 1 in 4 elements were out-of-bounds (see the 'ig_png' conversion functions documentation).
#> Warning: Variable 'sex': 1 in 4 elements were missing (`NA`).
#> [1]    NA 24.78    NA 43.02