Data from the Low birthweight Infant Feeding Exploration (LIFE) study

A dataset with anthropometric data for infants enrolled in the Low birthweight Infant Feeding Exploration (LIFE) study. The variables are as follows:

Format

A data frame with 10,026 rows and 24 variables:

infantid: Unique ID for each infant in the dataset. First number is site, second is facility ID (1--5) for that site. The last number denotes whether the infant is a singleton or twin.

motherid: Unique ID for each mother in the dataset. Same convention as infantid, but without sibling status.

site: Site ID, derived from first number of infantid (1 = Tanzania; 2 = Malawi; 3 = India-Karnataka; 4 = India-Odisha).

facility: Facility ID, derived from third number of infantid (1--5).

birthdate: Date of birth for a given infant.

visitweek: Chronological age in weeks when study visit occurred (0--26).

visitattend: Visit attendance status (1 = Attended; 2 = Missed).

visitdate: Date of visit as double.

visittime: Time of visit as character.

withdrawalrsn: Reason for withdrawal, if withdrawn (1 = Baby died; 2 = Mother moved away; 3 = Lost contact with mother; 4 = Mother withdrew consent; 5 = Mother died and new caregiver withdrew consent; 6 = Other; NA = No withdrawal).

deliverymode2: Mode of delivery for baby (1 = Vaginal delivery; 2 = Caesarean delivery).

birthcount: Number of babies born in delivery (1 = Singleton; 2 = Twins). Note: not all babies from twin births were enrolled in the study.

sibling: Siblings enrolled in the study (1--2)

sex: Sex of the infant (1 = Male; 2 = Female).

gestage: Best obstetric estimate of gestational age (132--311). Note: missing or implausible (<24 weeks) values excluded from analysis

preterm: Term status based on gestage (0 = Term (>=37 weeks gestage); 1 = Preterm (<37 weeks gestage).

sizega: Size for gestational age (1 = Small for GA (SGA); 2 = Appropriate for GA (AGA); 3 = Large for GA (LGA)).

LBWtype4: Low birthweight type (1 = Preterm + SGA; 2 = Preterm + AGA; 3 = Preterm + LGA; 4 = Term + SGA).

pma: Post-menstrual age in days (133--545)

age_days: Chronological age in days derived from gestage and pma (0--245).

meaninfwgt: Mean weight in g (1233.33--9905).

meaninflen: Mean length in cm (35.03--72.93).

meanhead: Mean head circumference in cm (23.20--45.73).

meanmuac: Mean mid-upper arm circumference (5.97--17.03).

References

Vesel L, Bellad RM, Manji K, Saidi F, Velasquez E, Sudfeld C, et al. Feeding practices and growth patterns of moderately low birthweight infants in resource-limited settings: results from a multisite, longitudinal observational study. BMJ Open 2023, 13(2):e067316. doi:10.1136/BMJOPEN-2022-067316

Examples

head(gigs::life6mo)
#>     infantid   motherid site facility  birthdate visitweek visitattend
#> 1 101-1002-1 101-1002-0    1        1 2019-10-09         0           1
#> 2 101-1002-1 101-1002-0    1        1 2019-10-09         1           1
#> 3 101-1002-1 101-1002-0    1        1 2019-10-09         2           1
#> 4 101-1002-1 101-1002-0    1        1 2019-10-09         4           1
#> 5 101-1002-1 101-1002-0    1        1 2019-10-09         6           1
#> 6 101-1002-1 101-1002-0    1        1 2019-10-09        10           1
#>    visitdate  visittime withdrawalrsn deliverymode2 birthcount sibling sex
#> 1 2019-10-10 10H 49M 0S            NA             1          1       1   F
#> 2 2019-10-16 10H 39M 0S            NA             1          1       1   F
#> 3 2019-10-23 11H 12M 0S            NA             1          1       1   F
#> 4 2019-11-08  15H 0M 0S            NA             1          1       1   F
#> 5 2019-11-20 15H 42M 0S            NA             1          1       1   F
#> 6 2019-12-19  10H 8M 0S            NA             1          1       1   F
#>   gestage preterm sizega LBWtype4 pma age_days meaninfwgt meaninflen meanhead
#> 1     132       1     NA       NA 133        1       2300   45.00000 30.53333
#> 2     132       1     NA       NA 139        7       2270   45.06667 30.53333
#> 3     132       1     NA       NA 146       14       2465   46.00000 31.16667
#> 4     132       1     NA       NA 162       30       2700   48.13333 33.16667
#> 5     132       1     NA       NA 174       42       3160   50.20000 34.00000
#> 6     132       1     NA       NA 203       71       4004   54.06667 36.03333
#>    meanmuac
#> 1  9.400000
#> 2  9.466667
#> 3  9.766667
#> 4  9.866666
#> 5 10.000000
#> 6 10.966667