Convert values to z-scores/centiles in the WHO Child Growth Standards

Convert values to z-scores/centiles in the WHO Child Growth Standards

Usage

who_gs_value2zscore(y, x, sex, acronym)

who_gs_wfa_value2zscore(weight_kg, age_days, sex)

who_gs_bfa_value2zscore(bmi, age_days, sex)

who_gs_lhfa_value2zscore(lenht_cm, age_days, sex)

who_gs_wfl_value2zscore(weight_kg, length_cm, sex)

who_gs_wfh_value2zscore(weight_kg, height_cm, sex)

who_gs_hcfa_value2zscore(headcirc_cm, age_days, sex)

who_gs_acfa_value2zscore(armcirc_cm, age_days, sex)

who_gs_ssfa_value2zscore(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2zscore(triceps_sf_mm, age_days, sex)

who_gs_value2centile(y, x, sex, acronym)

who_gs_wfa_value2centile(weight_kg, age_days, sex)

who_gs_bfa_value2centile(bmi, age_days, sex)

who_gs_lhfa_value2centile(lenht_cm, age_days, sex)

who_gs_wfl_value2centile(weight_kg, length_cm, sex)

who_gs_wfh_value2centile(weight_kg, height_cm, sex)

who_gs_hcfa_value2centile(headcirc_cm, age_days, sex)

who_gs_acfa_value2centile(armcirc_cm, age_days, sex)

who_gs_ssfa_value2centile(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2centile(triceps_sf_mm, age_days, sex)

Arguments

y

Numeric vector of length one or more with anthropometric measurement(s) to convert to centiles/z-scores. Units depend on which acronym(s) are in use.

x, age_days, length_cm, height_cm

Numeric vector of length one or more with x values. Elements in x or its standard-specific equivalents (age_days, length_cm, and height_cm) should have specific units and be between certain values depending on the standard in use (defined by acronym). These are:

Between 0 and 1856 days for "wfa", "bfa", "lhfa", and "hcfa".

Between 45 and 110 days for "wfl".

Between 65 and 120 days for "wfh".

Between 91 and 1856 days for "acfa", "ssfa", "tsfa".

By default, gigs will replace out-of-bounds elements in x with NA and warn you. This behaviour can be customised using the functions in gigs_options.

sex

Character vector of length one or more with sex(es), either "M" (male) or "F" (female). This argument is case-sensitive. By default, gigs will replace elements of sex which are not "M" or "F" with NA and warn you. This behaviour can be customised using the functions in gigs_options.

acronym

Character vector of length one or more denoting the WHO Child Growth standard(s) in use. Each element should be one of:

"wfa" (weight-for-age)

"bfa" (BMI-for-age)

"lhfa" (length/height-for-age)

"wfl" (weight-for-length)

"wfh" (weight-for-height)

"hcfa" (head circumference-for-age)

"acfa" (arm circumference-for-age)

"ssfa" (subscapular skinfold-for-age)

"tsfa" (triceps skinfold-for-age)

This argument is case-sensitive. By default, gigs will replace elements in acronym which are not one of the above values with NA and warn you. This behaviour can be customised using the functions in gigs_options. If all elements in acronym are not one of the above values, gigs will throw an error.

weight_kg

Numeric vector of length one or more with weight measurement(s) in kg.

bmi

Numeric vector of length one or more with body mass index measurement(s) in kg/m2.

lenht_cm

Numeric vector of length one or more with length/height measurement(s) in cm.

headcirc_cm

Numeric vector of length one or more with head circumference measurement(s) in cm.

armcirc_cm

Numeric vector of length one or more with arm circumference measurement(s) in cm.

subscap_sf_mm

Numeric vector of length one or more with subscapular skinfold measurement(s) in mm.

triceps_sf_mm

Numeric vector of length one or more with triceps skinfold measurement(s) in mm.

Value

Numeric vector of z-scores/centiles with length equal to the longest input vector.

Note

Input vectors are recycled by vctrs::vec_recycle_common(), and must adhere to the vctrs recycling rules.

References

de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. The WHO Multicentre Growth Reference Study: planning, study design, and methodology Food Nutr Bull. 2004, 25(1 Suppl):S15-26. doi:10.1177/15648265040251s104

World Health Organisation. WHO child growth standards: length/height-for-age, weight-for-age, weight-for-length, weight-for-height and body mass index-for-age: methods and development. Technical report, WHO, Geneva, 2006.

World Health Organisation. WHO child growth standards: head circumference-for-age, arm circumference-for-age, triceps skinfold-for-age and subscapular skinfold-for-age: methods and development. Technical report, WHO, Geneva, 2007.

Cole TJ. The LMS method for constructing normalized growth standards Eur J Clin Nutr. 1990, 44(1):45-60. PMID: 2354692

Examples

# Convert values to centiles
who_gs_value2centile(y = c(10.1, 10.2, 10.3),
                     x = 504:506,
                     sex = c("M", "F", "M"),
                     acronym = "wfa") |>
  round(digits = 2)
#> [1] 0.32 0.58 0.38

# Or values to z-scores
who_gs_value2zscore(y = c(10.1, 10.2, 10.3),
                    x = 504:506,
                    sex = c("M", "F", "M"),
                    acronym = "wfa") |>
  round(digits = 2)
#> [1] -0.47  0.21 -0.31

# Specify which standard to use with the acronym parameter...
who_gs_value2centile(y = c(73.9, 67.9, 71.4),
                     x = 299:301,
                     sex = c("M", "F", "M"),
                     acronym = "lhfa") |>
  round(digits = 2)
#> [1] 0.64 0.08 0.22

# ... or by using a standard-specific function
who_gs_lhfa_value2centile(lenht_cm = c(73.9, 67.9, 71.4),
                          age_days = 299:301,
                          sex = c("M", "F", "M")) |>
  round(digits = 2)
#> [1] 0.64 0.08 0.22

# Inputs are recycled to the input of the longest length
who_gs_ssfa_value2centile(subscap_sf_mm = 6.1,
                          age_days = seq(450, 650, by = 50),
                          sex = "M") |>
  round(digits = 2)
#> [1] 0.44 0.47 0.50 0.52 0.54

# Bad inputs will not stop the function but will instead produce `NA`s in the
# output - by default gigs will issue useful warnings
who_gs_wfh_value2zscore(weight_kg = c(7.4, 14.2, 21.8, 22.4),
                        height_cm = c(65, 95, 120, NA),
                        sex = c("M", "M", "F", NA)) |>
    round(digits = 2)
#> Warning: Variable 'height_cm': 1 in 4 elements were missing (`NA`).
#> Warning: Variable 'sex': 1 in 4 elements were missing (`NA`).
#> [1] -0.05  0.12 -0.47    NA