Convert z-scores/centiles to values in the WHO Child Growth Standards

Convert z-scores/centiles to values in the WHO Child Growth Standards

Usage

who_gs_zscore2value(z, x, sex, acronym)

who_gs_wfa_zscore2value(z, age_days, sex)

who_gs_bfa_zscore2value(z, age_days, sex)

who_gs_lhfa_zscore2value(z, age_days, sex)

who_gs_wfl_zscore2value(z, length_cm, sex)

who_gs_wfh_zscore2value(z, height_cm, sex)

who_gs_hcfa_zscore2value(z, age_days, sex)

who_gs_acfa_zscore2value(z, age_days, sex)

who_gs_ssfa_zscore2value(z, age_days, sex)

who_gs_tsfa_zscore2value(z, age_days, sex)

who_gs_centile2value(p, x, sex, acronym)

who_gs_wfa_centile2value(p, age_days, sex)

who_gs_bfa_centile2value(p, age_days, sex)

who_gs_lhfa_centile2value(p, age_days, sex)

who_gs_wfl_centile2value(p, length_cm, sex)

who_gs_wfh_centile2value(p, height_cm, sex)

who_gs_hcfa_centile2value(p, age_days, sex)

who_gs_acfa_centile2value(p, age_days, sex)

who_gs_ssfa_centile2value(p, age_days, sex)

who_gs_tsfa_centile2value(p, age_days, sex)

Arguments

z, p

Numeric vector of length one or more with centiles/z-scores to convert to values. For p, if an element of p is not between 0 and 1, gigs will replace it with NA and warn you. This behaviour can be customised using the functions in gigs_options.

x, age_days, length_cm, height_cm

Numeric vector of length one or more with x values. Elements in x or its standard-specific equivalents (age_days, length_cm, and height_cm) should have specific units and be between certain values depending on the standard in use (defined by acronym). These are:

Between 0 and 1856 days for "wfa", "bfa", "lhfa", and "hcfa".

Between 45 and 110 days for "wfl".

Between 65 and 120 days for "wfh".

Between 91 and 1856 days for "acfa", "ssfa", "tsfa".

By default, gigs will replace out-of-bounds elements in x with NA and warn you. This behaviour can be customised using the functions in gigs_options.

sex

Character vector of length one or more with sex(es), either "M" (male) or "F" (female). This argument is case-sensitive. By default, gigs will replace elements of sex which are not "M" or "F" with NA and warn you. This behaviour can be customised using the functions in gigs_options.

acronym

Character vector of length one or more denoting the WHO Child Growth standard(s) in use. Each element should be one of:

"wfa" (weight-for-age)

"bfa" (BMI-for-age)

"lhfa" (length/height-for-age)

"wfl" (weight-for-length)

"wfh" (weight-for-height)

"hcfa" (head circumference-for-age)

"acfa" (arm circumference-for-age)

"ssfa" (subscapular skinfold-for-age)

"tsfa" (triceps skinfold-for-age)

This argument is case-sensitive. By default, gigs will replace elements in acronym which are not one of the above values with NA and warn you. This behaviour can be customised using the functions in gigs_options. If all elements in acronym are not one of the above values, gigs will throw an error.

Value

Numeric vector of expected measurements with length equal to the longest input vector.

Note

Input vectors are recycled by vctrs::vec_recycle_common(), and must adhere to the vctrs recycling rules.

References

de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. The WHO Multicentre Growth Reference Study: planning, study design, and methodology Food Nutr Bull. 2004, 25(1 Suppl):S15-26. doi:10.1177/15648265040251s104

World Health Organisation. WHO child growth standards: length/height-for-age, weight-for-age, weight-for-length, weight-for-height and body mass index-for-age: methods and development. Technical report, WHO, Geneva, 2006.

World Health Organisation. WHO child growth standards: head circumference-for-age, arm circumference-for-age, triceps skinfold-for-age and subscapular skinfold-for-age: methods and development. Technical report, WHO, Geneva, 2007.

Cole TJ. The LMS method for constructing normalized growth standards Eur J Clin Nutr. 1990, 44(1):45-60. PMID: 2354692

Examples

# Convert centiles to values
who_gs_centile2value(p = c(0.25, 0.5, 0.75), # 25th, 50th, 70th centiles
                     x = 500:502,
                     sex = c("M", "F", "M"),
                     acronym = "wfa") |>
  round(digits = 2)
#> [1]  9.85  9.91 11.45

# Or z-scores to values
who_gs_zscore2value(z = qnorm(c(0.25, 0.5, 0.75)),
                    x = 500:502,
                    sex = c("M", "F", "M"),
                    acronym = "wfa") |>
  round(digits = 2)
#> [1]  9.85  9.91 11.45

# Specify which standard to use with the acronym parameter...
who_gs_zscore2value(z = -1:1,
                    x = 349:351,
                    sex = c("M", "M", "M"),
                    acronym = "lhfa") |>
  round(digits = 2)
#> [1] 72.76 75.15 77.54

# ... or by using a standard-specific function
who_gs_lhfa_zscore2value(z = -1:1,
                         age_days = 349:351,
                         sex = c("M", "M", "M")) |>
  round(digits = 2)
#> [1] 72.76 75.15 77.54

# Inputs are recycled to the input of the longest length
who_gs_ssfa_centile2value(p = seq(0.1, 0.9, by = 0.2),
                          age_days = 670,
                          sex = "M") |>
  round(digits = 2)
#> [1] 4.81 5.45 5.97 6.58 7.66

# Bad inputs will not stop the function but will instead return `NA`, and by
# default print useful warnings
who_gs_wfh_centile2value(p = c(0.2, 0.4, 0.6, 15),
                         height_cm = c(60, 85, 105, 120),
                         sex = c("M", "M", "M", NA)) |>
  round(digits = 2)
#> Warning: Variable 'p': 1 in 4 elements were not between 0 and 1.
#> Warning: Variable 'height_cm': 1 in 4 elements were out-of-bounds (see the 'who_gs' conversion functions documentation).
#> Warning: Variable 'sex': 1 in 4 elements were missing (`NA`).
#> [1]    NA 11.43 17.21    NA