Classify small vulnerable newborns according to the INTERGROWTH-21st weight-for-gestational age standard

A small vulnerable newborn (SVN) can be preterm (born too soon), small-for-gestational age (born too small), or within both of these categories. This function uses weight and gestational age to categorise newborns according to their SVN type, for use in downstream analyses.

Usage

classify_svn(weight_kg, gest_days, sex)

Arguments

weight_kg

Numeric vector of length one or more with weight value(s) in kg.

gest_days

Numeric vector of length one or more with gestational age(s) at birth in days between 168 and 300. By default, gigs will replace out-of-bounds elements in gest_days with NA and warn you. This behaviour can be customised using the functions in gigs_options.

sex

Character vector of length one or more with sex(es), either "M" (male) or "F" (female). This argument is case-sensitive. By default, gigs will replace elements of sex which are not "M" or "F" with NA and warn you. This behaviour can be customised using the functions in gigs_options.

Value

An object of class factor with the same length as the longest input vector, containing small vulnerable newborn classifications. Its levels are c("Preterm SGA", "Preterm AGA", "Preterm LGA", "Term SGA", "Term AGA", "Term LGA").

Note

Input vectors are recycled by vctrs::vec_recycle_common(), and must adhere to the vctrs recycling rules.

References

Lawn JE, Ohuma EO, Bradley E, Idueta LS, Hazel E, Okwaraji YB et al. Small babies, big risks: global estimates of prevalence and mortality for vulnerable newborns to accelerate change and improve counting. The Lancet 2023, 401(10389):1707-1719. doi:10.1016/S0140-6736(23)00522-6

See also

Examples

classify_svn(
  weight_kg = c(1.5, 2.6, 2.6, 3.5),
  gest_days = c(235, 257, 275, 295),
  sex = c("F", "M", "F", "M")
)
#> [1] Preterm SGA Preterm AGA Term SGA    Term AGA   
#> Levels: Preterm SGA Preterm AGA Preterm LGA Term SGA Term AGA Term LGA