Convert anthropometric measures to WHO Growth Standards z-scores

Convert anthropometric measures to WHO Growth Standards z-scores

Usage

who_gs_value2zscore(y, x, sex, acronym)

who_gs_wfa_value2zscore(weight_kg, age_days, sex)

who_gs_bfa_value2zscore(bmi, age_days, sex)

who_gs_lhfa_value2zscore(lenht_cm, age_days, sex)

who_gs_wfl_value2zscore(weight_kg, length_cm, sex)

who_gs_wfh_value2zscore(weight_kg, height_cm, sex)

who_gs_hcfa_value2zscore(headcirc_cm, age_days, sex)

who_gs_acfa_value2zscore(armcirc_cm, age_days, sex)

who_gs_ssfa_value2zscore(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2zscore(triceps_sf_mm, age_days, sex)

who_gs_value2percentile(y, x, sex, acronym)

who_gs_wfa_value2percentile(weight_kg, age_days, sex)

who_gs_bfa_value2percentile(bmi, age_days, sex)

who_gs_lhfa_value2percentile(lenht_cm, age_days, sex)

who_gs_wfl_value2percentile(weight_kg, length_cm, sex)

who_gs_wfh_value2percentile(weight_kg, height_cm, sex)

who_gs_hcfa_value2percentile(headcirc_cm, age_days, sex)

who_gs_acfa_value2percentile(armcirc_cm, age_days, sex)

who_gs_ssfa_value2percentile(subscap_sf_mm, age_days, sex)

who_gs_tsfa_value2percentile(triceps_sf_mm, age_days, sex)

Arguments

y

Y value(s) for growth standard.

x

X value(s) for growth standard.

sex

Sex(es), either "M" (male) or "F" (female).

acronym

Valid acronym for WHO Growth Standards datasets: one of "wfa" (weight-for-age), "bfa" (bmi-for-age), "lhfa" (length/height-for-age), "wfl" (weight-for-length), or "wfh" (weight-for-height), "hcfa" (head circumference-for-age), "acfa" (arm circumference-for-age), "ssfa" (subscapular skinfold-for-age), or "tsfa" (triceps skinfold-for-age),

weight_kg

Weight measurement(s) in kg.

age_days

Age since birth in days.

bmi

Body mass index measurement(s).

lenht_cm

Length/height measurement(s) in cm.

length_cm

Recumbent length measurement(s) in cm.

height_cm

Standing height measurement(s) in cm.

headcirc_cm

Head circumference measurement(s) in cm.

armcirc_cm

Arm circumference measurement(s) in cm.

subscap_sf_mm

Subscapular skinfold measurement(s) in mm.

triceps_sf_mm

Triceps skinfold measurement(s) in mm.

References

de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. The WHO Multicentre Growth Reference Study: planning, study design, and methodology Food Nutr Bull. 2004, 25(1 Suppl):S15-26. doi: 10.1177/15648265040251s104

de Onis M et al. (2006) 'COMPUTATION OF CENTILES AND Z-SCORES FOR LENGTH/HEIGHT-FOR-AGE, WEIGHT-FOR-AGE, WEIGHT-FOR-LENGTH, WEIGHT-FOR-HEIGHT AND BMI-FOR-AGE' in World Health Organization (ed.) WHO Child Growth Standards: length/height-for-age, weight-for-age, weight-for-length, weight-for-height and body mass index-for-age: methods and development. Geneva: World Health Organization, pp. 301-304.

Cole TJ The LMS method for constructing normalized growth standards Eur J Clin Nutr. 1990, 44(1):45-60. PMID: 2354692

Examples

# Convert values to percentiles
who_gs_value2percentile(y = 10.1, x = 505, sex = "M", acronym = "wfa") |>
  round(digits = 2)
#> [1] 0.32

# Or values to z-scores
who_gs_value2zscore(y = 10.1, x = 505, sex = "M", acronym = "wfa") |>
  round(digits = 2)
#> [1] -0.48

# Specify which standard to use with the acronym parameter...
who_gs_value2percentile(y = 75.2, x = 300, sex = "M", acronym = "lhfa") |>
  round(digits = 2)
#> [1] 0.82

# ... or by using a standard-specific function
who_gs_lhfa_value2percentile(lenht_cm = 75.2, age_days = 300, sex = "M") |>
  round(digits = 2)
#> [1] 0.82

# Inputs are recycled to the input of the longest length
who_gs_ssfa_value2percentile(subscap_sf_mm = 6.1,
                             age_days = seq(450, 650, by = 50),
                             sex = "M") |>
  round(digits = 2)
#> [1] 0.44 0.47 0.50 0.52 0.54

# Bad inputs will not stop the function but will instead return NA - here 121 cm
# height_cm is incompatible with the WHO Growth Standard for weight-for-height
who_gs_wfh_value2zscore(weight_kg = c(7.4, 14.2, 21.8, 22.4),
                        height_cm = c(65, 95, 120, 121),
                        sex = "M") |>
  round(digits = 2)
#> [1] -0.05  0.12 -0.27    NA