Convert z-scores/percentiles to values in the WHO Child Growth Standards

Convert z-scores/percentiles to values in the WHO Child Growth Standards

Usage

who_gs_zscore2value(z, x, sex, acronym)

who_gs_wfa_zscore2value(z, age_days, sex)

who_gs_bfa_zscore2value(z, age_days, sex)

who_gs_lhfa_zscore2value(z, age_days, sex)

who_gs_wfl_zscore2value(z, length_cm, sex)

who_gs_wfh_zscore2value(z, height_cm, sex)

who_gs_hcfa_zscore2value(z, age_days, sex)

who_gs_acfa_zscore2value(z, age_days, sex)

who_gs_ssfa_zscore2value(z, age_days, sex)

who_gs_tsfa_zscore2value(z, age_days, sex)

who_gs_percentile2value(p, x, sex, acronym)

who_gs_wfa_percentile2value(p, age_days, sex)

who_gs_bfa_percentile2value(p, age_days, sex)

who_gs_lhfa_percentile2value(p, age_days, sex)

who_gs_wfl_percentile2value(p, length_cm, sex)

who_gs_wfh_percentile2value(p, height_cm, sex)

who_gs_hcfa_percentile2value(p, age_days, sex)

who_gs_acfa_percentile2value(p, age_days, sex)

who_gs_ssfa_percentile2value(p, age_days, sex)

who_gs_tsfa_percentile2value(p, age_days, sex)

Arguments

z, p

Z-score(s)/percentiles to convert to a value/values.

x, age_days, length_cm, height_cm

The x value at which to convert a value to a z-score/percentile. Must be within bounds of available x values for given acronym. The standard-specific versions of each function specify x, either age_days (age since birth in days), length_cm (recumbent length measurement(s) in cm) or height_cm (standing height measurement(s) in cm).

sex

Sex(es), either "M" (male) or "F" (female).

acronym

Valid acronym(s) for the WHO Child Growth Standards: either "wfa" (weight-for-age), "bfa" (BMI-for-age), "lhfa" (length/height-for-age), "wfl" (weight-for-length), or "wfh" (weight-for-height), "hcfa" (head circumference-for-age), "acfa" (arm circumference-for-age), "ssfa" (subscapular skinfold-for-age), or "tsfa" (triceps skinfold-for-age).

Value

Expected measurements for each combination of z-score/percentile, x variable, sex, and acronym provided to the function.

References

de Onis M, Garza C, Victora CG, Onyango AW, Frongillo EA, Martines J. The WHO Multicentre Growth Reference Study: planning, study design, and methodology Food Nutr Bull. 2004, 25(1 Suppl):S15-26. doi: 10.1177/15648265040251s104

World Health Organisation. WHO child growth standards: length/height-for-age, weight-for-age, weight-for-length, weight-for-height and body mass index-for-age: methods and development. Technical report, WHO, Geneva, 2006.

World Health Organisation. WHO child growth standards: head circumference-for-age, arm circumference-for-age, triceps skinfold-for-age and subscapular skinfold-for-age: methods and development. Technical report, WHO, Geneva, 2007.

Cole TJ. The LMS method for constructing normalized growth standards Eur J Clin Nutr. 1990, 44(1):45-60. PMID: 2354692

Examples

# Convert percentiles to values
p <- 0.25 # 25th percentile
who_gs_percentile2value(p = p, x = 501, sex = "M", acronym = "wfa") |>
  round(digits = 2)
#> [1] 9.86

# Or z-scores to values
z <- qnorm(p) # Z-score equivalent of 25th percentile
who_gs_zscore2value(z = z, x = 501, sex = "M", acronym = "wfa") |>
  round(digits = 2)
#> [1] 9.86

# Specify which standard to use with the acronym parameter...
who_gs_zscore2value(z = z, x = 300, sex = "M", acronym = "lhfa") |>
  round(digits = 2)
#> [1] 71.56

# ... or by using a standard-specific function
who_gs_lhfa_zscore2value(z = z, age_days = 300, sex = "M") |>
  round(digits = 2)
#> [1] 71.56

# Inputs are recycled to the input of the longest length
who_gs_ssfa_percentile2value(p = seq(0.1, 0.9, by = 0.2),
                             age_days = 670,
                             sex = "M") |>
  round(digits = 2)
#> [1] 4.81 5.45 5.97 6.58 7.66

# Bad inputs will not stop the function but will instead return NA - here 60
# cm for height_cm is incompatible with the WHO Growth Standard for
# weight-for-height
who_gs_wfh_zscore2value(z = 0,
                        height_cm = c(60, 85, 105, 120),
                        sex = "M") |>
  round(digits = 2)
#> [1]    NA 11.67 16.85 22.35